requirements.txt
fastapi
uvicorn[standard]
sse_starlette
firebase-admin
google-cloud-firestore
python-dotenv